requirements.txt
# === 核心数据处理 ===
pandas>=2.0.0
numpy>=1.23.0
pyyaml>=6.0

# === 可视化 ===
matplotlib>=3.6.0

# === 配置管理（.env）===
python-dotenv>=1.0.0

# === 开发/测试工具 ===
pytest>=7.0.0

# === 股票/金融数据接口 ===
akshare>=1.10.0       # 免费的 A 股、美股、港股数据接口
tushare>=1.2.0        # 专业级数据平台（需要 token）

# === 策略评估与机器学习（可选）===
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=3.3.0

# === MySQL 数据库支持 ===
mysql-connector-python>=8.0.27  # MySQL 连接器，用于数据库连接


# === 可选：因子分析与量化工具（后续扩展）===
# alphalens-reloaded     # 多因子分析工具（Quantopian Fork）

# === Neovim 补全支持 coc.nvim 时需要 ===
# (仅开发机器使用) nodejs 版本由 coc 管理

# === Jupyter Notebook 环境（可选）===
notebook
jupyterlab

# === Parquet 支持 ===
pyarrow>=14.0.0
# 或者使用 fastparquet
# fastparquet>=2024.0.0
